Ship class completed
Completed the Ship class with the following attributes: name, size,
hits_left, coords_A, coords_D, and is_sunk. Also have a hit method
that records a hit for the ship. This class is used in conjunction
with the board and player class to make a functioning sea battle game.